+++ src/Form/CreditCardDetailsType.php
@@ -1,27 +1,28 @@
 <?php
 
-namespace App\Form\order;
 
-use App\Entity\PaymentDetails;
+namespace App\Form;
+
+
+use App\Entity\CreditCardDetails;
 use Symfony\Component\Form\AbstractType;
 use Symfony\Component\Form\FormBuilderInterface;
 use Symfony\Component\OptionsResolver\OptionsResolver;
 
-class PaymentDetailsType extends AbstractType
+class CreditCardDetailsType extends AbstractType
 {
     public function buildForm(FormBuilderInterface $builder, array $options)
     {
         $builder
-            ->add('type')
-            ->add('currency')
-            ->add('details')
-        ;
+            ->add('code')
+            ->add('cvv')
+            ->add('expiresAt');
     }
 
     public function configureOptions(OptionsResolver $resolver)
     {
         $resolver->setDefaults([
-            'data_class' => PaymentDetails::class,
+            'data_class' => CreditCardDetails::class,
         ]);
     }
 }
