--- src/Form/order/OrderType.php
@@ -1,30 +1,0 @@
-<?php
-
-namespace App\Form\order;
-
-use App\Entity\Order;
-use Symfony\Component\Form\AbstractType;
-use Symfony\Component\Form\FormBuilderInterface;
-use Symfony\Component\OptionsResolver\OptionsResolver;
-
-class OrderType extends AbstractType
-{
-    public function buildForm(FormBuilderInterface $builder, array $options)
-    {
-        $builder
-            ->add('items')
-            ->add('paymentDetails')
-            ->add('address')
-            ->add('total')
-            ->add('status')
-            ->add('createdAt')
-        ;
-    }
-
-    public function configureOptions(OptionsResolver $resolver)
-    {
-        $resolver->setDefaults([
-            'data_class' => Order::class,
-        ]);
-    }
-}
