--- templates/address/_delete_form.html.twig
@@ -1,4 +1,0 @@
-<form method="post" action="{{ path('address_delete', {'id': address.id}) }}" onsubmit="return confirm('Are you sure you want to delete this item?');">
-    <input type="hidden" name="_token" value="{{ csrf_token('delete' ~ address.id) }}">
-    <button class="btn">Delete</button>
-</form>
