--- templates/address/edit.html.twig
@@ -1,13 +1,0 @@
-{% extends 'base.html.twig' %}
-
-{% block title %}Edit Address{% endblock %}
-
-{% block body %}
-    <h1>Edit Address</h1>
-
-    {{ include('address/_form.html.twig', {'button_label': 'Update'}) }}
-
-    <a href="{{ path('address_index') }}">back to list</a>
-
-    {{ include('address/_delete_form.html.twig') }}
-{% endblock %}
