--- templates/address/show.html.twig
@@ -1,42 +1,0 @@
-{% extends 'base.html.twig' %}
-
-{% block title %}Address{% endblock %}
-
-{% block body %}
-    <h1>Address</h1>
-
-    <table class="table">
-        <tbody>
-            <tr>
-                <th>Id</th>
-                <td>{{ address.id }}</td>
-            </tr>
-            <tr>
-                <th>Country</th>
-                <td>{{ address.country }}</td>
-            </tr>
-            <tr>
-                <th>State</th>
-                <td>{{ address.state }}</td>
-            </tr>
-            <tr>
-                <th>City</th>
-                <td>{{ address.city }}</td>
-            </tr>
-            <tr>
-                <th>Address</th>
-                <td>{{ address.address }}</td>
-            </tr>
-            <tr>
-                <th>PhoneNumber</th>
-                <td>{{ address.phoneNumber }}</td>
-            </tr>
-        </tbody>
-    </table>
-
-    <a href="{{ path('address_index') }}">back to list</a>
-
-    <a href="{{ path('address_edit', {'id': address.id}) }}">edit</a>
-
-    {{ include('address/_delete_form.html.twig') }}
-{% endblock %}
