--- templates/credit_card_details/_delete_form.html.twig
@@ -1,4 +1,0 @@
-<form method="post" action="{{ path('credit_card_details_delete', {'id': credit_card_detail.id}) }}" onsubmit="return confirm('Are you sure you want to delete this item?');">
-    <input type="hidden" name="_token" value="{{ csrf_token('delete' ~ credit_card_detail.id) }}">
-    <button class="btn">Delete</button>
-</form>
