--- templates/credit_card_details/edit.html.twig
@@ -1,13 +1,0 @@
-{% extends 'base.html.twig' %}
-
-{% block title %}Edit CreditCardDetails{% endblock %}
-
-{% block body %}
-    <h1>Edit CreditCardDetails</h1>
-
-    {{ include('credit_card_details/_form.html.twig', {'button_label': 'Update'}) }}
-
-    <a href="{{ path('credit_card_details_index') }}">back to list</a>
-
-    {{ include('credit_card_details/_delete_form.html.twig') }}
-{% endblock %}
