--- templates/credit_card_details/index.html.twig
@@ -1,39 +1,0 @@
-{% extends 'base.html.twig' %}
-
-{% block title %}CreditCardDetails index{% endblock %}
-
-{% block body %}
-    <h1>CreditCardDetails index</h1>
-
-    <table class="table">
-        <thead>
-            <tr>
-                <th>Id</th>
-                <th>Code</th>
-                <th>Cvv</th>
-                <th>ExpiresAt</th>
-                <th>actions</th>
-            </tr>
-        </thead>
-        <tbody>
-        {% for credit_card_detail in credit_card_details %}
-            <tr>
-                <td>{{ credit_card_detail.id }}</td>
-                <td>{{ credit_card_detail.code }}</td>
-                <td>{{ credit_card_detail.cvv }}</td>
-                <td>{{ credit_card_detail.expiresAt ? credit_card_detail.expiresAt|date('Y-m-d') : '' }}</td>
-                <td>
-                    <a href="{{ path('credit_card_details_show', {'id': credit_card_detail.id}) }}">show</a>
-                    <a href="{{ path('credit_card_details_edit', {'id': credit_card_detail.id}) }}">edit</a>
-                </td>
-            </tr>
-        {% else %}
-            <tr>
-                <td colspan="5">no records found</td>
-            </tr>
-        {% endfor %}
-        </tbody>
-    </table>
-
-    <a href="{{ path('credit_card_details_new') }}">Create new</a>
-{% endblock %}
