--- templates/credit_card_details/show.html.twig
@@ -1,34 +1,0 @@
-{% extends 'base.html.twig' %}
-
-{% block title %}CreditCardDetails{% endblock %}
-
-{% block body %}
-    <h1>CreditCardDetails</h1>
-
-    <table class="table">
-        <tbody>
-            <tr>
-                <th>Id</th>
-                <td>{{ credit_card_detail.id }}</td>
-            </tr>
-            <tr>
-                <th>Code</th>
-                <td>{{ credit_card_detail.code }}</td>
-            </tr>
-            <tr>
-                <th>Cvv</th>
-                <td>{{ credit_card_detail.cvv }}</td>
-            </tr>
-            <tr>
-                <th>ExpiresAt</th>
-                <td>{{ credit_card_detail.expiresAt ? credit_card_detail.expiresAt|date('Y-m-d') : '' }}</td>
-            </tr>
-        </tbody>
-    </table>
-
-    <a href="{{ path('credit_card_details_index') }}">back to list</a>
-
-    <a href="{{ path('credit_card_details_edit', {'id': credit_card_detail.id}) }}">edit</a>
-
-    {{ include('credit_card_details/_delete_form.html.twig') }}
-{% endblock %}
