--- templates/order/edit.html.twig
@@ -1,13 +1,0 @@
-{% extends 'base.html.twig' %}
-
-{% block title %}Edit Order{% endblock %}
-
-{% block body %}
-    <h1>Edit Order</h1>
-
-    {{ include('order/_form.html.twig', {'button_label': 'Update'}) }}
-
-    <a href="{{ path('order_index') }}">back to list</a>
-
-    {{ include('order/_delete_form.html.twig') }}
-{% endblock %}
