--- templates/order/index.html.twig
@@ -1,45 +1,0 @@
-{% extends 'base.html.twig' %}
-
-{% block title %}Order index{% endblock %}
-
-{% block body %}
-    <h1>Order index</h1>
-
-    <table class="table">
-        <thead>
-            <tr>
-                <th>Id</th>
-                <th>Items</th>
-                <th>PaymentDetails</th>
-                <th>Address</th>
-                <th>Total</th>
-                <th>Status</th>
-                <th>CreatedAt</th>
-                <th>actions</th>
-            </tr>
-        </thead>
-        <tbody>
-        {% for order in orders %}
-            <tr>
-                <td>{{ order.id }}</td>
-                <td>{{ order.items ? order.items|join(', ') : '' }}</td>
-                <td>{{ order.paymentDetails }}</td>
-                <td>{{ order.address }}</td>
-                <td>{{ order.total }}</td>
-                <td>{{ order.status ? order.status|join(', ') : '' }}</td>
-                <td>{{ order.createdAt ? order.createdAt|date('Y-m-d H:i:s') : '' }}</td>
-                <td>
-                    <a href="{{ path('order_show', {'id': order.id}) }}">show</a>
-                    <a href="{{ path('order_edit', {'id': order.id}) }}">edit</a>
-                </td>
-            </tr>
-        {% else %}
-            <tr>
-                <td colspan="8">no records found</td>
-            </tr>
-        {% endfor %}
-        </tbody>
-    </table>
-
-    <a href="{{ path('order_new') }}">Create new</a>
-{% endblock %}
