--- templates/order/new.html.twig
@@ -1,11 +1,0 @@
-{% extends 'base.html.twig' %}
-
-{% block title %}New Order{% endblock %}
-
-{% block body %}
-    <h1>Create new Order</h1>
-
-    {{ include('order/_form.html.twig') }}
-
-    <a href="{{ path('order_index') }}">back to list</a>
-{% endblock %}
