--- templates/order/show.html.twig
@@ -1,46 +1,0 @@
-{% extends 'base.html.twig' %}
-
-{% block title %}Order{% endblock %}
-
-{% block body %}
-    <h1>Order</h1>
-
-    <table class="table">
-        <tbody>
-            <tr>
-                <th>Id</th>
-                <td>{{ order.id }}</td>
-            </tr>
-            <tr>
-                <th>Items</th>
-                <td>{{ order.items ? order.items|join(', ') : '' }}</td>
-            </tr>
-            <tr>
-                <th>PaymentDetails</th>
-                <td>{{ order.paymentDetails }}</td>
-            </tr>
-            <tr>
-                <th>Address</th>
-                <td>{{ order.address }}</td>
-            </tr>
-            <tr>
-                <th>Total</th>
-                <td>{{ order.total }}</td>
-            </tr>
-            <tr>
-                <th>Status</th>
-                <td>{{ order.status ? order.status|join(', ') : '' }}</td>
-            </tr>
-            <tr>
-                <th>CreatedAt</th>
-                <td>{{ order.createdAt ? order.createdAt|date('Y-m-d H:i:s') : '' }}</td>
-            </tr>
-        </tbody>
-    </table>
-
-    <a href="{{ path('order_index') }}">back to list</a>
-
-    <a href="{{ path('order_edit', {'id': order.id}) }}">edit</a>
-
-    {{ include('order/_delete_form.html.twig') }}
-{% endblock %}
