--- templates/order_item/_delete_form.html.twig
@@ -1,4 +1,0 @@
-<form method="post" action="{{ path('order_item_delete', {'id': order_item.id}) }}" onsubmit="return confirm('Are you sure you want to delete this item?');">
-    <input type="hidden" name="_token" value="{{ csrf_token('delete' ~ order_item.id) }}">
-    <button class="btn">Delete</button>
-</form>
