--- templates/order_item/edit.html.twig
@@ -1,13 +1,0 @@
-{% extends 'base.html.twig' %}
-
-{% block title %}Edit OrderItem{% endblock %}
-
-{% block body %}
-    <h1>Edit OrderItem</h1>
-
-    {{ include('order_item/_form.html.twig', {'button_label': 'Update'}) }}
-
-    <a href="{{ path('order_item_index') }}">back to list</a>
-
-    {{ include('order_item/_delete_form.html.twig') }}
-{% endblock %}
