--- templates/order_item/index.html.twig
@@ -1,39 +1,0 @@
-{% extends 'base.html.twig' %}
-
-{% block title %}OrderItem index{% endblock %}
-
-{% block body %}
-    <h1>OrderItem index</h1>
-
-    <table class="table">
-        <thead>
-            <tr>
-                <th>Id</th>
-                <th>ProductCode</th>
-                <th>Amount</th>
-                <th>Price</th>
-                <th>actions</th>
-            </tr>
-        </thead>
-        <tbody>
-        {% for order_item in order_items %}
-            <tr>
-                <td>{{ order_item.id }}</td>
-                <td>{{ order_item.productCode }}</td>
-                <td>{{ order_item.amount }}</td>
-                <td>{{ order_item.price }}</td>
-                <td>
-                    <a href="{{ path('order_item_show', {'id': order_item.id}) }}">show</a>
-                    <a href="{{ path('order_item_edit', {'id': order_item.id}) }}">edit</a>
-                </td>
-            </tr>
-        {% else %}
-            <tr>
-                <td colspan="5">no records found</td>
-            </tr>
-        {% endfor %}
-        </tbody>
-    </table>
-
-    <a href="{{ path('order_item_new') }}">Create new</a>
-{% endblock %}
