--- templates/order_item/new.html.twig
@@ -1,11 +1,0 @@
-{% extends 'base.html.twig' %}
-
-{% block title %}New OrderItem{% endblock %}
-
-{% block body %}
-    <h1>Create new OrderItem</h1>
-
-    {{ include('order_item/_form.html.twig') }}
-
-    <a href="{{ path('order_item_index') }}">back to list</a>
-{% endblock %}
