--- templates/order_item/show.html.twig
@@ -1,34 +1,0 @@
-{% extends 'base.html.twig' %}
-
-{% block title %}OrderItem{% endblock %}
-
-{% block body %}
-    <h1>OrderItem</h1>
-
-    <table class="table">
-        <tbody>
-            <tr>
-                <th>Id</th>
-                <td>{{ order_item.id }}</td>
-            </tr>
-            <tr>
-                <th>ProductCode</th>
-                <td>{{ order_item.productCode }}</td>
-            </tr>
-            <tr>
-                <th>Amount</th>
-                <td>{{ order_item.amount }}</td>
-            </tr>
-            <tr>
-                <th>Price</th>
-                <td>{{ order_item.price }}</td>
-            </tr>
-        </tbody>
-    </table>
-
-    <a href="{{ path('order_item_index') }}">back to list</a>
-
-    <a href="{{ path('order_item_edit', {'id': order_item.id}) }}">edit</a>
-
-    {{ include('order_item/_delete_form.html.twig') }}
-{% endblock %}
