--- templates/payment_details/edit.html.twig
@@ -1,13 +1,0 @@
-{% extends 'base.html.twig' %}
-
-{% block title %}Edit PaymentDetails{% endblock %}
-
-{% block body %}
-    <h1>Edit PaymentDetails</h1>
-
-    {{ include('payment_details/_form.html.twig', {'button_label': 'Update'}) }}
-
-    <a href="{{ path('payment_details_index') }}">back to list</a>
-
-    {{ include('payment_details/_delete_form.html.twig') }}
-{% endblock %}
