--- templates/payment_details/index.html.twig
@@ -1,39 +1,0 @@
-{% extends 'base.html.twig' %}
-
-{% block title %}PaymentDetails index{% endblock %}
-
-{% block body %}
-    <h1>PaymentDetails index</h1>
-
-    <table class="table">
-        <thead>
-            <tr>
-                <th>Id</th>
-                <th>Type</th>
-                <th>Currency</th>
-                <th>Details</th>
-                <th>actions</th>
-            </tr>
-        </thead>
-        <tbody>
-        {% for payment_detail in payment_details %}
-            <tr>
-                <td>{{ payment_detail.id }}</td>
-                <td>{{ payment_detail.type }}</td>
-                <td>{{ payment_detail.currency }}</td>
-                <td>{{ payment_detail.details }}</td>
-                <td>
-                    <a href="{{ path('payment_details_show', {'id': payment_detail.id}) }}">show</a>
-                    <a href="{{ path('payment_details_edit', {'id': payment_detail.id}) }}">edit</a>
-                </td>
-            </tr>
-        {% else %}
-            <tr>
-                <td colspan="5">no records found</td>
-            </tr>
-        {% endfor %}
-        </tbody>
-    </table>
-
-    <a href="{{ path('payment_details_new') }}">Create new</a>
-{% endblock %}
