--- templates/payment_details/new.html.twig
@@ -1,11 +1,0 @@
-{% extends 'base.html.twig' %}
-
-{% block title %}New PaymentDetails{% endblock %}
-
-{% block body %}
-    <h1>Create new PaymentDetails</h1>
-
-    {{ include('payment_details/_form.html.twig') }}
-
-    <a href="{{ path('payment_details_index') }}">back to list</a>
-{% endblock %}
